中机

+++ zhongjiJian/UpdateAllPersonJobBySomeone.jsp
@@ -3,6 +3,7 @@
 <%@ page import="weaver.conn.RecordSet" %>
 <%@ page import="weaver.general.BaseBean" %>
 <%@ page import="java.util.ArrayList" %>
+<%@ page import="java.util.HashMap" %>
 <%@ page import="java.util.List" %>
 <%@ page import="java.util.Map" %>
 <%@ page language="java" contentType="text/html; charset=UTF-8" %>
@@ -10,8 +11,6 @@
 <%
     // 更新 人员信息查询中，业绩信息tab
     BaseBean baseBean = new BaseBean();
-    RecordSet recordSet = new RecordSet();
-    RecordSet updateSet = new RecordSet();
     String xm = request.getParameter("xm");
     String selectSql = "SELECT\n" +
             "\ta.id,\n" +
@@ -59,29 +58,58 @@
             baseBean.writeLog("xm为空，不作操作");
             return;
         }
+        RecordSet recordSet = new RecordSet();
+        RecordSet updateSet = new RecordSet();
 
-        synchronized (ConnUtil.class) {
-            updateSet.executeUpdate("delete from uf_perjob where ryid = ?", xm);
-            recordSet.executeQuery(selectSql, xm);
-            List<String> list = new ArrayList<>();
-            while (recordSet.next()) {
-                String id = recordSet.getString("id"); // 自定义人员表id uf_ryzcxxb
-                String xmid = recordSet.getString("xmid"); // 项目名称
-                for (Map.Entry<String, String> entry : colMap.entrySet()) {
-                    String curVal = recordSet.getString(entry.getKey());
-                    if (id.equals(curVal)) {
-                        list.add(entry.getValue());
-                    }
-                }
-
-                String joinerStr = String.join(", ", list);
-                // 插入或更新到 人员 - 岗位表
-                updateSet.executeUpdate(updateSql,
-                        id, xmid, joinerStr);
-                list.clear();
-            }
-            baseBean.writeLog("更新一人建模人员信息岗End");
+        // 查询当前人已有的业绩数据条数
+        List<String> idList = new ArrayList<>();
+        recordSet.executeQuery("select id from uf_perjob where ryid = ?", xm);
+        int oldCounts = recordSet.getCounts();
+        baseBean.writeLog("已有数据条数： " + oldCounts);
+        while (recordSet.next()) {
+            idList.add(recordSet.getString("id"));
         }
+
+        recordSet.executeQuery(selectSql, xm);
+        int newCounts = recordSet.getCounts(); // 本次将要操作的数据条数
+        baseBean.writeLog("本次操作数据条数： " + newCounts);
+
+        Map<String, String> map = new HashMap<>();
+        List<String> list = new ArrayList<>();
+        while (recordSet.next()) {
+            String xmid = recordSet.getString("xmid"); // 项目名称
+            for (Map.Entry<String, String> entry : colMap.entrySet()) {
+                String curVal = recordSet.getString(entry.getKey());
+                if (xm.equals(curVal)) {
+                    list.add(entry.getValue());
+                }
+            }
+            String joinerStr = String.join(", ", list);
+            map.put(xmid, joinerStr);
+            list.clear();
+        }
+
+        if (oldCounts == newCounts) {
+            // 总条数未变，执行更新
+            int i = 0;
+            for (Map.Entry<String, String> entry : map.entrySet()) {
+                updateSet.executeUpdate("update uf_perjob set xmmc = ?, gwmc = ? where id = ?",
+                        entry.getKey(), entry.getValue(), idList.get(i));
+                i++;
+            }
+        } else {
+            synchronized (ConnUtil.class) {
+                // 全部删除，新插入
+                updateSet.executeUpdate("delete from uf_perjob where ryid = ?", xm);
+                // 插入或更新到 人员 - 岗位表
+                for (Map.Entry<String, String> entry : map.entrySet()) {
+                    updateSet.executeUpdate(updateSql,
+                            xm, entry.getKey(), entry.getValue());
+                }
+            }
+
+        }
+        baseBean.writeLog("更新一人建模人员信息岗End");
 
         response.sendRedirect("/spa/cube/index.html#/main/cube/search?customid=151&hidetop=1&xm=" + xm);
     } catch (Exception e) {
